Organize S3 Classes in Amazon Provider (#20167)
* Task: Organize S3 Classes in Amazon Provider
