+++ airflow/providers/amazon/aws/example_dags/example_s3_bucket_tagging.py
@@ -18,8 +18,9 @@ import os
 from datetime import datetime
 
 from airflow.models.dag import DAG
-from airflow.providers.amazon.aws.operators.s3_bucket import S3CreateBucketOperator, S3DeleteBucketOperator
-from airflow.providers.amazon.aws.operators.s3_bucket_tagging import (
+from airflow.providers.amazon.aws.operators.s3 import (
+    S3CreateBucketOperator,
+    S3DeleteBucketOperator,
     S3DeleteBucketTaggingOperator,
     S3GetBucketTaggingOperator,
     S3PutBucketTaggingOperator,
