+++ airflow/providers/amazon/aws/example_dags/example_salesforce_to_s3.py
@@ -22,8 +22,7 @@ data and upload to an S3 bucket.
 from datetime import datetime
 
 from airflow import DAG
-from airflow.providers.amazon.aws.operators.s3_copy_object import S3CopyObjectOperator
-from airflow.providers.amazon.aws.operators.s3_delete_objects import S3DeleteObjectsOperator
+from airflow.providers.amazon.aws.operators.s3 import S3CopyObjectOperator, S3DeleteObjectsOperator
 from airflow.providers.amazon.aws.transfers.salesforce_to_s3 import SalesforceToS3Operator
 
 BASE_PATH = "salesforce/customers"
