+++ dev/provider_packages/prepare_provider_packages.py
@@ -2141,6 +2141,8 @@ KNOWN_DEPRECATED_DIRECT_IMPORTS: Set[str] = {
     "This module is deprecated. Please use `airflow.providers.amazon.aws.operators.step_function`.",
     'This module is deprecated. Please use `airflow.providers.amazon.aws.operators.ec2`.',
     'This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.ec2`.',
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.s3`.",
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.operators.s3`.",
 }
