+++ tests/deprecated_classes.py
@@ -1056,16 +1056,20 @@ OPERATORS = [
         'airflow.contrib.operators.emr_terminate_job_flow_operator.EmrTerminateJobFlowOperator',
     ),
     (
-        'airflow.providers.amazon.aws.operators.s3_copy_object.S3CopyObjectOperator',
+        'airflow.providers.amazon.aws.operators.s3.S3CopyObjectOperator',
         'airflow.contrib.operators.s3_copy_object_operator.S3CopyObjectOperator',
     ),
     (
-        'airflow.providers.amazon.aws.operators.s3_delete_objects.S3DeleteObjectsOperator',
+        'airflow.providers.amazon.aws.operators.s3.S3DeleteObjectsOperator',
         'airflow.contrib.operators.s3_delete_objects_operator.S3DeleteObjectsOperator',
     ),
     (
-        'airflow.providers.amazon.aws.operators.s3_list.S3ListOperator',
+        'airflow.providers.amazon.aws.operators.s3.S3ListOperator',
         'airflow.contrib.operators.s3_list_operator.S3ListOperator',
+    ),
+    (
+        'airflow.providers.amazon.aws.operators.s3.S3FileTransformOperator',
+        'airflow.operators.s3_file_transform_operator.S3FileTransformOperator',
     ),
     (
         'airflow.providers.amazon.aws.operators.sagemaker_base.SageMakerBaseOperator',
@@ -1351,6 +1355,46 @@ OPERATORS = [
         "airflow.providers.amazon.aws.operators.ec2.EC2StopInstanceOperator",
         "airflow.providers.amazon.aws.operators.ec2_stop_instance.EC2StopInstanceOperator",
     ),
+    (
+        "airflow.providers.amazon.aws.operators.s3.S3FileTransformOperator",
+        "airflow.providers.amazon.aws.operators.s3_file_transform.S3FileTransformOperator",
+    ),
+    (
+        "airflow.providers.amazon.aws.operators.s3.S3ListOperator",
+        "airflow.providers.amazon.aws.operators.s3_list.S3ListOperator",
+    ),
+    (
+        "airflow.providers.amazon.aws.operators.s3.S3CreateBucketOperator",
+        "airflow.providers.amazon.aws.operators.s3_bucket.S3CreateBucketOperator",
+    ),
+    (
+        "airflow.providers.amazon.aws.operators.s3.S3DeleteBucketOperator",
+        "airflow.providers.amazon.aws.operators.s3_bucket.S3DeleteBucketOperator",
+    ),
+    (
+        "airflow.providers.amazon.aws.operators.s3.S3GetBucketTaggingOperator",
+        "airflow.providers.amazon.aws.operators.s3_bucket_tagging.S3GetBucketTaggingOperator",
+    ),
+    (
+        "airflow.providers.amazon.aws.operators.s3.S3PutBucketTaggingOperator",
+        "airflow.providers.amazon.aws.operators.s3_bucket_tagging.S3PutBucketTaggingOperator",
+    ),
+    (
+        "airflow.providers.amazon.aws.operators.s3.S3DeleteBucketTaggingOperator",
+        "airflow.providers.amazon.aws.operators.s3_bucket_tagging.S3DeleteBucketTaggingOperator",
+    ),
+    (
+        "airflow.providers.amazon.aws.operators.s3.S3CopyObjectOperator",
+        "airflow.providers.amazon.aws.operators.s3_copy_object.S3CopyObjectOperator",
+    ),
+    (
+        "airflow.providers.amazon.aws.operators.s3.S3DeleteObjectsOperator",
+        "airflow.providers.amazon.aws.operators.s3_delete_objects.S3DeleteObjectsOperator",
+    ),
+    (
+        "airflow.providers.amazon.aws.operators.s3.S3ListPrefixesOperator",
+        "airflow.providers.amazon.aws.operators.s3_list_prefixes.S3ListPrefixesOperator",
+    ),
 ]
 
 SECRETS = [
@@ -1515,10 +1559,6 @@ SENSORS = [
         'airflow.contrib.sensors.sagemaker_tuning_sensor.SageMakerTuningSensor',
     ),
     (
-        'airflow.providers.amazon.aws.operators.s3_file_transform.S3FileTransformOperator',
-        'airflow.operators.s3_file_transform_operator.S3FileTransformOperator',
-    ),
-    (
         'airflow.providers.amazon.aws.operators.step_function.StepFunctionStartExecutionOperator',
         'airflow.providers.amazon.aws.operators.step_function_start_execution'
         '.StepFunctionStartExecutionOperator',
@@ -1529,11 +1569,11 @@ SENSORS = [
         '.StepFunctionGetExecutionOutputOperator',
     ),
     (
-        'airflow.providers.amazon.aws.sensors.s3_key.S3KeySensor',
+        'airflow.providers.amazon.aws.sensors.s3.S3KeySensor',
         'airflow.sensors.s3_key_sensor.S3KeySensor',
     ),
     (
-        'airflow.providers.amazon.aws.sensors.s3_prefix.S3PrefixSensor',
+        'airflow.providers.amazon.aws.sensors.s3.S3PrefixSensor',
         'airflow.sensors.s3_prefix_sensor.S3PrefixSensor',
     ),
     (
@@ -1611,6 +1651,22 @@ SENSORS = [
     (
         'airflow.providers.amazon.aws.sensors.ec2.EC2InstanceStateSensor',
         'airflow.providers.amazon.aws.sensors.ec2_instance_state.EC2InstanceStateSensor',
+    ),
+    (
+        'airflow.providers.amazon.aws.sensors.s3.S3KeySensor',
+        'airflow.providers.amazon.aws.sensors.s3_key.S3KeySensor',
+    ),
+    (
+        'airflow.providers.amazon.aws.sensors.s3.S3KeySizeSensor',
+        'airflow.providers.amazon.aws.sensors.s3_key.S3KeySizeSensor',
+    ),
+    (
+        'airflow.providers.amazon.aws.sensors.s3.S3KeysUnchangedSensor',
+        'airflow.providers.amazon.aws.sensors.s3_keys_unchanged.S3KeysUnchangedSensor',
+    ),
+    (
+        "airflow.providers.amazon.aws.sensors.s3.S3PrefixSensor",
+        "airflow.providers.amazon.aws.sensors.s3_prefix.S3PrefixSensor",
     ),
 ]
