+++ tests/providers/amazon/aws/operators/test_s3_bucket.py
@@ -22,7 +22,7 @@ from unittest import mock
 from moto import mock_s3
 
 from airflow.providers.amazon.aws.hooks.s3 import S3Hook
-from airflow.providers.amazon.aws.operators.s3_bucket import S3CreateBucketOperator, S3DeleteBucketOperator
+from airflow.providers.amazon.aws.operators.s3 import S3CreateBucketOperator, S3DeleteBucketOperator
 
 BUCKET_NAME = os.environ.get("BUCKET_NAME", "test-airflow-bucket")
 TASK_ID = os.environ.get("TASK_ID", "test-s3-operator")
