+++ tests/providers/amazon/aws/operators/test_s3_bucket_tagging.py
@@ -22,7 +22,7 @@ from unittest import mock
 from moto import mock_s3
 
 from airflow.providers.amazon.aws.hooks.s3 import S3Hook
-from airflow.providers.amazon.aws.operators.s3_bucket_tagging import (
+from airflow.providers.amazon.aws.operators.s3 import (
     S3DeleteBucketTaggingOperator,
     S3GetBucketTaggingOperator,
     S3PutBucketTaggingOperator,
