+++ tests/providers/amazon/aws/operators/test_s3_list_prefixes.py
@@ -19,7 +19,7 @@
 import unittest
 from unittest import mock
 
-from airflow.providers.amazon.aws.operators.s3_list_prefixes import S3ListPrefixesOperator
+from airflow.providers.amazon.aws.operators.s3 import S3ListPrefixesOperator
 
 TASK_ID = 'test-s3-list-prefixes-operator'
 BUCKET = 'test-bucket'
@@ -29,7 +29,7 @@ MOCK_SUBFOLDERS = ['test/']
 
 
 class TestS3ListOperator(unittest.TestCase):
-    @mock.patch('airflow.providers.amazon.aws.operators.s3_list_prefixes.S3Hook')
+    @mock.patch('airflow.providers.amazon.aws.operators.s3.S3Hook')
     def test_execute(self, mock_hook):
 
         mock_hook.return_value.list_prefixes.return_value = MOCK_SUBFOLDERS
