+++ tests/providers/amazon/aws/sensors/test_s3_keys_unchanged.py
@@ -24,7 +24,7 @@ from freezegun import freeze_time
 from parameterized import parameterized
 
 from airflow.models.dag import DAG, AirflowException
-from airflow.providers.amazon.aws.sensors.s3_keys_unchanged import S3KeysUnchangedSensor
+from airflow.providers.amazon.aws.sensors.s3 import S3KeysUnchangedSensor
 
 TEST_DAG_ID = 'unit_tests_aws_sensor'
 DEFAULT_DATE = datetime(2015, 1, 1)
@@ -100,7 +100,7 @@ class TestS3KeysUnchangedSensor(TestCase):
         assert self.sensor.inactivity_seconds == inactivity_periods[2]
 
     @freeze_time(DEFAULT_DATE, auto_tick_seconds=10)
-    @mock.patch('airflow.providers.amazon.aws.sensors.s3_keys_unchanged.S3Hook')
+    @mock.patch('airflow.providers.amazon.aws.sensors.s3.S3Hook')
     def test_poke_succeeds_on_upload_complete(self, mock_hook):
         mock_hook.return_value.list_keys.return_value = {'a'}
         assert not self.sensor.poke(dict())
